doing things on first loginw

+++ install.sh
@@ -16,15 +16,21 @@ if [[ \$- =~ i ]] && [[ -z "\$TMUX" ]] && [[ -n "\$SSH_TTY" ]]; then
 fi
 EOL
 
-# Installing some things
-sudo apt-get update && sudo apt-get -q upgrade
-sudo apt-get -q vim tmux curl git psmisc
+# Updating .profile
+cat >> ~/.profile <<EOL
+if [ ! -f "\$HOME/.bash-git-prompt/gitprompt.sh" ]; then
+  echo "Initial setup - please wait"
+  sudo apt-get update && sudo apt-get -q upgrade
+  sudo apt-get -q vim tmux curl git psmisc
+  /usr/bin/git clone https://github.com/magicmonty/bash-git-prompt.git ~/.bash-git-prompt --depth=1
+  mv ~/dotfiles/.vimrc ~/.vimrc
+  curl -fLo ~/.vim/autoload/plug.vim --create-dirs \
+      https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+  /usr/bin/vim +PlugInstall +qall
+fi
+if [[ \$- =~ i ]] && [[ -z "\$TMUX" ]] && [[ -n "\$SSH_TTY" ]]; then
+  tmux attach-session -t ssh_tmux || tmux new-session -s ssh_tmux
+fi
+EOL
 
-# Moving things
-/usr/bin/git clone https://github.com/magicmonty/bash-git-prompt.git ~/.bash-git-prompt --depth=1
 
-# Configuring VIM
-mv ~/dotfiles/.vimrc ~/.vimrc
-curl -fLo ~/.vim/autoload/plug.vim --create-dirs \
-    https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
-/usr/bin/vim +PlugInstall +qall
